Adds basic structure for multilanguage
This commit adds the JSON with the translated defaultView strings and
the JS file needed to translate.

In addition, it adds the corresponding buttons in the main view for the
language change and creates a cookie to store the language. It adapts
the Translate.js file to also translate the placeholders too.