Updates basic structure
This commit modifies DefaultDAO to support integer inserts and to
insert as NULL the empty values. If the insertion/edition returns
some error, it is shown in a toast.

In addition, this commit modifies all ids to be AUTO_INCREMENT with
the exception of academic course, wich continues as varchar with format:
(start_year/end_year).

This commit also allows to translate message shown in toasts.