Fixes translation and adds support to translate dynamic texts
This commit fixes the translation because it didn't load well before.
In addition, adds support to translate dynamic texts translating the
whole sentence except the words entered between two %. To add this
translations in translations.json file, it is necessary to enter %c
in the place where they will be introduced.