Refactorizes all the code to use entities instead of ids
This commit changes all entities to use the entity they depend on
instead of their id.

It also adds pagination and search to all showAll views. It also
creates a dependency control by deleting that shows the number of
entities in each table that have dependencies with that entity.

In addition, some refactorigs were made in the code, especially in
the controllers.